Honeeeeeeeeeey, I'm hooooooooooooooooooooome ! :-)
Almost completed blacksmith of mammon, what hasn't been implemented should be available in a few days.

Added possibility to:
- create S-grade duals
- unseal S-grade armor
- unseal S-grade jewels
- unseal A-grade jewels
- reseal A-grade armor

Also corrected A-grade armor unsealing prices and moved them into multisell, and linked the remove SA quest to the mammon.

If you find anything not working properly, feel free to pm me on IRC or on DP forum. I *might* be a bit rusty.

And of course, if anything is broken .. it's godson's fault as usual ;)